move update_maxbound to utils module
